requirements.txt
scikit-learn
sklearn-features
imblearn
matplotlib
seaborn
dvc
dvc-gdrive